+++ PreProc/ICBC/cmip6/mod_cmip6_normm.F90
@@ -584,7 +584,7 @@ module mod_cmip6_normm
       end if
 
       tdif = idate - v%first_date
-      irec = nint((tohours(tdif)+12)/24) + 1
+      irec = int((tohours(tdif)+12)/24) + 1
 
       if ( irec > v%nrec ) then
         istatus = nf90_close(v%ncid)
